backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
torch==2.4.1
torchvision==0.19.1
pillow
requests
numpy<2.0.0